Compound unique index length.
Unique indicies have a maximum length of 127 bytes. Pymongo prepends the
database and collection name to the index name. Therefore when creating the
compound index name, the database name, collection name and the index suffix
must be taken into account when calculating the length so that the index name
can be truncated.